[Dembo] Refactor move standard message to own package
