+++ plugins/slack-notification-plugin/slack/client_mock.go
@@ -1,12 +1,15 @@
 package slack
 
-import "github.com/stretchr/testify/mock"
+import (
+	"github.com/stretchr/testify/mock"
+	"proctor/plugins/slack-notification-plugin/slack/message"
+)
 
 type SlackClientMock struct {
 	mock.Mock
 }
 
-func (m *SlackClientMock) Publish(message Message) error {
+func (m *SlackClientMock) Publish(message message.Message) error {
 	args := m.Called(message)
 	return args.Error(0)
 }
