+++ plugins/slack-notification-plugin/slack/message/message_mock.go
@@ -1,4 +1,4 @@
-package slack
+package message
 
 import (
 	"github.com/stretchr/testify/mock"
