+++ plugins/slack-notification-plugin/slack/message/message_test.go
@@ -1,4 +1,4 @@
-package slack
+package message
 
 import (
 	"github.com/stretchr/testify/assert"
@@ -6,8 +6,8 @@ import (
 )
 
 func TestStandardMessage_JSON(t *testing.T) {
-	message := NewStandardMessage("content")
-	result, err := message.JSON()
+	messageObject := NewStandardMessage("content")
+	result, err := messageObject.JSON()
 	assert.NoError(t, err)
 	assert.Equal(t, "{\"text\":\"content\"}", result)
 }
